# speed up arithmetic