--- a/libeantic/e-antic/renf_elem_class.hpp
+++ b/libeantic/e-antic/renf_elem_class.hpp
@@ -17,7 +17,7 @@
 #include <gmpxx.h>
 #include <boost/operators.hpp>
 #include <flint/fmpq.h>
-#include <memory>
+#include <boost/smart_ptr/intrusive_ptr.hpp>
 #include <vector>
 
 #include "e-antic.h"
@@ -53,34 +53,34 @@ public:
     // The zero element in k; note that all overloads that take the field as a
     // parameter hold a non-owning reference to the field, i.e., the element is
     // only valid while that reference is.
-    explicit renf_elem_class(std::shared_ptr<const renf_class> k);
+    explicit renf_elem_class(const renf_class& k);
     // An integer in the field k
-    renf_elem_class(std::shared_ptr<const renf_class> k, const mpz_class &);
+    renf_elem_class(const renf_class& k, const mpz_class &);
     // A rational in the field k
-    renf_elem_class(std::shared_ptr<const renf_class> k, const mpq_class &);
+    renf_elem_class(const renf_class& k, const mpq_class &);
     // An integer in the field k
-    renf_elem_class(std::shared_ptr<const renf_class> k, const fmpz_t);
+    renf_elem_class(const renf_class& k, const fmpz_t);
     // A rational in the field k
-    renf_elem_class(std::shared_ptr<const renf_class> k, const fmpq_t);
+    renf_elem_class(const renf_class& k, const fmpq_t);
     // An integer in the field k
-    renf_elem_class(std::shared_ptr<const renf_class> k, const int);
-    renf_elem_class(std::shared_ptr<const renf_class> k, const unsigned int);
-    renf_elem_class(std::shared_ptr<const renf_class> k, const long);
-    renf_elem_class(std::shared_ptr<const renf_class> k, const unsigned long);
-    renf_elem_class(std::shared_ptr<const renf_class> k, const long long);
-    renf_elem_class(std::shared_ptr<const renf_class> k, const unsigned long long);
+    renf_elem_class(const renf_class& k, int);
+    renf_elem_class(const renf_class& k, unsigned int);
+    renf_elem_class(const renf_class& k, long);
+    renf_elem_class(const renf_class& k, unsigned long);
+    renf_elem_class(const renf_class& k, long long);
+    renf_elem_class(const renf_class& k, unsigned long long);
     // Coerce number field element to the field k. Only implemented in trivial cases.
-    renf_elem_class(std::shared_ptr<const renf_class> k, const renf_elem_class&);
+    renf_elem_class(const renf_class& k, const renf_elem_class&);
     // Parse the string into an element in the field k
-    renf_elem_class(std::shared_ptr<const renf_class> k, const std::string &);
+    renf_elem_class(const renf_class& k, const std::string &);
     // The element Σc_i·α^i where α is the generator of the field k; the number
     // of coefficients must not exceed the degree of the field.
-    renf_elem_class(std::shared_ptr<const renf_class> k, const std::vector<int> &);
-    renf_elem_class(std::shared_ptr<const renf_class> k, const std::vector<unsigned int> &);
-    renf_elem_class(std::shared_ptr<const renf_class> k, const std::vector<long> &);
-    renf_elem_class(std::shared_ptr<const renf_class> k, const std::vector<unsigned long> &);
-    renf_elem_class(std::shared_ptr<const renf_class> k, const std::vector<mpz_class> &);
-    renf_elem_class(std::shared_ptr<const renf_class> k, const std::vector<mpq_class> &);
+    renf_elem_class(const renf_class& k, const std::vector<int> &);
+    renf_elem_class(const renf_class& k, const std::vector<unsigned int> &);
+    renf_elem_class(const renf_class& k, const std::vector<long> &);
+    renf_elem_class(const renf_class& k, const std::vector<unsigned long> &);
+    renf_elem_class(const renf_class& k, const std::vector<mpz_class> &);
+    renf_elem_class(const renf_class& k, const std::vector<mpq_class> &);
 
     ~renf_elem_class() noexcept;
 
@@ -96,9 +96,6 @@ public:
     renf_elem_class & operator=(const fmpq_t);
     renf_elem_class & operator=(const renf_elem_class &);
     renf_elem_class & operator=(renf_elem_class &&) noexcept;
-
-    renf_elem_class & reset(std::shared_ptr<const renf_class>&&);
-    renf_elem_class & reset(const std::shared_ptr<const renf_class>&);
 
     // containing number field
     const renf_class& parent() const { return *nf; }
@@ -232,7 +229,7 @@ public:
 
 private:
     // The parent number field.
-    std::shared_ptr<const renf_class> nf;
+    boost::intrusive_ptr<const renf_class> nf;
     // The underlying element.
     // We need mutability as calls might need to refine the precision of
     // the stored embedding.
@@ -240,6 +237,8 @@ private:
 
     // Serialization, see cereal.hpp
     friend cereal::access;
+
+    LIBEANTIC_API friend void swap(renf_elem_class& lhs, renf_elem_class& rhs) noexcept;
 };
 
 // overloads for global functions