--- a/libeantic/src/renf_elem/add_fmpq.c
+++ b/libeantic/src/renf_elem/add_fmpq.c
@@ -9,7 +9,7 @@
     (at your option) any later version.  See <http://www.gnu.org/licenses/>.
 */
 
-#include "../e-antic/renf_elem.h"
+#include "../../e-antic/renf_elem.h"
 
 static void arb_add_fmpq(arb_t a, const arb_t b, const fmpq_t c, slong prec)
 {