--- a/libeantic/src/renf_elem/clear.c
+++ b/libeantic/src/renf_elem/clear.c
@@ -17,5 +17,3 @@ void renf_elem_clear(renf_elem_t a, const renf_t nf)
     nf_elem_clear(a->elem, nf->nf);
     arb_clear(a->emb);
 }
-
-