--- a/libeantic/src/renf_elem/clear.c
+++ b/libeantic/src/renf_elem/clear.c
@@ -10,7 +10,7 @@
 */
 
 
-#include "../e-antic/renf_elem.h"
+#include "../../e-antic/renf_elem.h"
 
 void renf_elem_clear(renf_elem_t a, const renf_t nf)
 {