--- a/libeantic/src/renf_elem/init.c
+++ b/libeantic/src/renf_elem/init.c
@@ -10,7 +10,7 @@
 */
 
 
-#include "../e-antic/renf_elem.h"
+#include "../../e-antic/renf_elem.h"
 
 void renf_elem_init(renf_elem_t a, const renf_t nf)
 {