--- a/libeantic/srcxx/renf_elem_class.cpp
+++ b/libeantic/srcxx/renf_elem_class.cpp
@@ -84,9 +84,7 @@ void coerce(renf_elem_class& self, const renf_class& parent)
         fmpz_t value;
         fmpz_init_set(value, renf_elem_get_fmpz(self.renf_elem_t(), parent.renf_t()));
 
-        self = parent.zero();
-
-        renf_elem_set_fmpz(self.renf_elem_t(), value, self.parent().renf_t());
+        self = value;
 
         fmpz_clear(value);
     }