--- a/pyeantic/src/pyeantic/__init__.py
+++ b/pyeantic/src/pyeantic/__init__.py
@@ -6,8 +6,8 @@ Python Wrappers for E-ANTIC
 ######################################################################
 #  This file is part of e-antic.
 #
-#        Copyright (C) 2019 Vincent Delecroix
-#        Copyright (C) 2019 Julian Rüth
+#        Copyright (C)      2019 Vincent Delecroix
+#                      2019-2021 Julian Rüth
 #
 #  e-antic is free software: you can redistribute it and/or modify
 #  it under the terms of the GNU Lesser General Public License as published by