--- a/pyeantic/src/pyeantic/real_embedded_number_field.py
+++ b/pyeantic/src/pyeantic/real_embedded_number_field.py
@@ -10,7 +10,7 @@ required for classical geometry.
 #  This file is part of e-antic.
 #
 #        Copyright (C)      2019 Vincent Delecroix
-#                      2019-2020 Julian Rüth
+#                      2019-2021 Julian Rüth
 #
 #  e-antic is free software: you can redistribute it and/or modify
 #  it under the terms of the GNU Lesser General Public License as published by