@saraedum 2021-03-04
@videlec, let's discuss this later today.

The only interface change is a `renf_elem_class::reset(renf_class)` method that sets the number field of an element. It's a somewhat weird operation that zeros the element but changes the field the element it is contained in.