@videlec 2021-03-04
Sorted by ratio we have the very bads
```
TrivialDivision<int>/4                                         66.7 ns        120 ns          0.55
TrivialDivision<long>/4                                        66.7 ns        121 ns          0.55
TrivialDivision<long long>/4                                   84.5 ns        136 ns          0.62
```
and the very goods
```
TrivialReverseAddition/4                                        352 ns        172 ns          2.04
TrivialReverseSubtraction/4                                     345 ns        168 ns          2.05
TrivialAssignment<long long>/1                                 49.0 ns       23.4 ns          2.09
TrivialAssignment<mpz_class>/1                                 49.6 ns       21.4 ns          2.31
TrivialReverseMultiplication/2                                  314 ns        135 ns          2.32
TrivialReverseDivision/4                                        311 ns        130 ns          2.39
TrivialSubtraction<renf_elem_class>/4                           335 ns        133 ns          2.51
TrivialAddition<renf_elem_class>/4                              333 ns        128 ns          2.60
TrivialReverseSubtraction/2                                     288 ns        109 ns          2.64
TrivialAssignment<int>/1                                       41.3 ns       15.4 ns          2.68
TrivialReverseAddition/2                                        287 ns        104 ns          2.75
TrivialAssignment<long>/1                                      41.3 ns       14.4 ns          2.86
TrivialSubtraction<renf_elem_class>/2                           281 ns       50.7 ns          5.54
TrivialAddition<renf_elem_class>/2                              280 ns       49.6 ns          5.64
TrivialReverseDivision/2                                        313 ns       50.3 ns          6.22
TrivialMultiplication<renf_elem_class>/4                        391 ns       60.4 ns          6.47
TrivialMultiplication<renf_elem_class>/2                        323 ns       45.2 ns          7.14
TrivialDivision<renf_elem_class>/4                              575 ns       73.4 ns          7.83
TrivialDivision<renf_elem_class>/2                              984 ns       76.1 ns         12.93
```